add: ログアウトできるようにする
