+++ app/components/Dialog/index.tsx
@@ -3,15 +3,14 @@ import { Dialog as HeadlessDialog, Transition } from '@headlessui/react'
 
 type Props = {
   open: boolean
-  onCancel: () => void
-  onSubmit: () => void
+  onClose: () => void
   children: React.ReactNode
 }
 
-export const Dialog = ({ open, onCancel, onSubmit, children }: Props) => {
+export const Dialog = ({ open, onClose, children }: Props) => {
   return (
     <Transition.Root show={open} as={Fragment}>
-      <HeadlessDialog as="div" className="relative z-10" onClose={onCancel}>
+      <HeadlessDialog as="div" className="relative z-10" onClose={onClose}>
         <Transition.Child
           as={Fragment}
           enter="ease-out duration-300"
@@ -37,22 +36,6 @@ export const Dialog = ({ open, onCancel, onSubmit, children }: Props) => {
             >
               <HeadlessDialog.Panel className="relative overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg">
                 {children}
-                <div className="bg-base px-4 py-3 sm:flex sm:flex-row-reverse sm:px-6">
-                  <button
-                    type="button"
-                    className="inline-flex w-full justify-center rounded-md bg-red-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-red-500 sm:ml-3 sm:w-auto"
-                    onClick={onSubmit}
-                  >
-                    Deactivate
-                  </button>
-                  <button
-                    type="button"
-                    className="mt-3 inline-flex w-full justify-center rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50 sm:mt-0 sm:w-auto"
-                    onClick={onCancel}
-                  >
-                    Cancel
-                  </button>
-                </div>
               </HeadlessDialog.Panel>
             </Transition.Child>
           </div>
